testeUnitario/testngTutorial/src/main/java/br/com/gilmario/model/EmployeeDetails.java
package br.com.gilmario.model;

/**
 * Created by gilmario on 16/02/17.
 */
public class EmployeeDetails {
    private String name;
    private double MonthlySalary;
    private int age;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public double getMonthlySalary() {
        return MonthlySalary;
    }

    public void setMonthlySalary(double monthlySalary) {
        MonthlySalary = monthlySalary;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }
}
